creation of fault logs and synchronized start
fault logs are created by inserting a "man in the middle"
the transcriber plugin inserts itself between the server and the fault
manager. Thus he overhears all tics, failed nodes and the node names in
the beginning. For this mode, the option '-n <num>' was introduced which
sets an upper limit for the number of tics that will be simulated.

a state was added to the nodes that shows the current state with respect
to the fault management (e.g. state ready != no fault occured).
The server starts, once all nodes have indicated they are ready, if the
flag '-s' was given.